Refactor themes
