+++ lib/features/settings/theme_item.dart
@@ -1,0 +1,10 @@
+import 'package:flutter/material.dart';
+
+class ThemeItem extends StatelessWidget {
+  const ThemeItem({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return const Placeholder();
+  }
+}
