+++ lib/theme/brown/brown_theme.dart
@@ -1,0 +1,309 @@
+import 'package:flutter/material.dart';
+
+class BrownTheme {
+  static ColorScheme lightScheme() {
+    return const ColorScheme(
+      brightness: Brightness.light,
+      primary: Color(0xff845317),
+      surfaceTint: Color(0xff845317),
+      onPrimary: Color(0xffffffff),
+      primaryContainer: Color(0xffffdcbc),
+      onPrimaryContainer: Color(0xff683d00),
+      secondary: Color(0xff725a42),
+      onSecondary: Color(0xffffffff),
+      secondaryContainer: Color(0xfffeddbe),
+      onSecondaryContainer: Color(0xff58432c),
+      tertiary: Color(0xff57633b),
+      onTertiary: Color(0xffffffff),
+      tertiaryContainer: Color(0xffdae8b5),
+      onTertiaryContainer: Color(0xff3f4b25),
+      error: Color(0xffba1a1a),
+      onError: Color(0xffffffff),
+      errorContainer: Color(0xffffdad6),
+      onErrorContainer: Color(0xff93000a),
+      surface: Color(0xfffff8f4),
+      onSurface: Color(0xff211a14),
+      onSurfaceVariant: Color(0xff50453a),
+      outline: Color(0xff837568),
+      outlineVariant: Color(0xffd5c3b5),
+      shadow: Color(0xff000000),
+      scrim: Color(0xff000000),
+      inverseSurface: Color(0xff372f28),
+      inversePrimary: Color(0xfffbb974),
+      primaryFixed: Color(0xffffdcbc),
+      onPrimaryFixed: Color(0xff2c1700),
+      primaryFixedDim: Color(0xfffbb974),
+      onPrimaryFixedVariant: Color(0xff683d00),
+      secondaryFixed: Color(0xfffeddbe),
+      onSecondaryFixed: Color(0xff291805),
+      secondaryFixedDim: Color(0xffe1c1a3),
+      onSecondaryFixedVariant: Color(0xff58432c),
+      tertiaryFixed: Color(0xffdae8b5),
+      onTertiaryFixed: Color(0xff151f01),
+      tertiaryFixedDim: Color(0xffbecc9b),
+      onTertiaryFixedVariant: Color(0xff3f4b25),
+      surfaceDim: Color(0xffe6d8cd),
+      surfaceBright: Color(0xfffff8f4),
+      surfaceContainerLowest: Color(0xffffffff),
+      surfaceContainerLow: Color(0xfffff1e7),
+      surfaceContainer: Color(0xfffaebe0),
+      surfaceContainerHigh: Color(0xfff4e6da),
+      surfaceContainerHighest: Color(0xffeee0d5),
+    );
+  }
+
+  static ColorScheme lightMediumContrastScheme() {
+    return const ColorScheme(
+      brightness: Brightness.light,
+      primary: Color(0xff512e00),
+      surfaceTint: Color(0xff845317),
+      onPrimary: Color(0xffffffff),
+      primaryContainer: Color(0xff966224),
+      onPrimaryContainer: Color(0xffffffff),
+      secondary: Color(0xff46321d),
+      onSecondary: Color(0xffffffff),
+      secondaryContainer: Color(0xff82684f),
+      onSecondaryContainer: Color(0xffffffff),
+      tertiary: Color(0xff2f3a16),
+      onTertiary: Color(0xffffffff),
+      tertiaryContainer: Color(0xff657248),
+      onTertiaryContainer: Color(0xffffffff),
+      error: Color(0xff740006),
+      onError: Color(0xffffffff),
+      errorContainer: Color(0xffcf2c27),
+      onErrorContainer: Color(0xffffffff),
+      surface: Color(0xfffff8f4),
+      onSurface: Color(0xff16100a),
+      onSurfaceVariant: Color(0xff3f342a),
+      outline: Color(0xff5d5045),
+      outlineVariant: Color(0xff786b5f),
+      shadow: Color(0xff000000),
+      scrim: Color(0xff000000),
+      inverseSurface: Color(0xff372f28),
+      inversePrimary: Color(0xfffbb974),
+      primaryFixed: Color(0xff966224),
+      onPrimaryFixed: Color(0xffffffff),
+      primaryFixedDim: Color(0xff794a0c),
+      onPrimaryFixedVariant: Color(0xffffffff),
+      secondaryFixed: Color(0xff82684f),
+      onSecondaryFixed: Color(0xffffffff),
+      secondaryFixedDim: Color(0xff685039),
+      onSecondaryFixedVariant: Color(0xffffffff),
+      tertiaryFixed: Color(0xff657248),
+      onTertiaryFixed: Color(0xffffffff),
+      tertiaryFixedDim: Color(0xff4d5932),
+      onTertiaryFixedVariant: Color(0xffffffff),
+      surfaceDim: Color(0xffd2c4b9),
+      surfaceBright: Color(0xfffff8f4),
+      surfaceContainerLowest: Color(0xffffffff),
+      surfaceContainerLow: Color(0xfffff1e7),
+      surfaceContainer: Color(0xfff4e6da),
+      surfaceContainerHigh: Color(0xffe9dacf),
+      surfaceContainerHighest: Color(0xffddcfc4),
+    );
+  }
+
+  static ColorScheme lightHighContrastScheme() {
+    return const ColorScheme(
+      brightness: Brightness.light,
+      primary: Color(0xff432500),
+      surfaceTint: Color(0xff845317),
+      onPrimary: Color(0xffffffff),
+      primaryContainer: Color(0xff6b3f00),
+      onPrimaryContainer: Color(0xffffffff),
+      secondary: Color(0xff3b2814),
+      onSecondary: Color(0xffffffff),
+      secondaryContainer: Color(0xff5b452e),
+      onSecondaryContainer: Color(0xffffffff),
+      tertiary: Color(0xff25300d),
+      onTertiary: Color(0xffffffff),
+      tertiaryContainer: Color(0xff424d27),
+      onTertiaryContainer: Color(0xffffffff),
+      error: Color(0xff600004),
+      onError: Color(0xffffffff),
+      errorContainer: Color(0xff98000a),
+      onErrorContainer: Color(0xffffffff),
+      surface: Color(0xfffff8f4),
+      onSurface: Color(0xff000000),
+      onSurfaceVariant: Color(0xff000000),
+      outline: Color(0xff342a20),
+      outlineVariant: Color(0xff53473c),
+      shadow: Color(0xff000000),
+      scrim: Color(0xff000000),
+      inverseSurface: Color(0xff372f28),
+      inversePrimary: Color(0xfffbb974),
+      primaryFixed: Color(0xff6b3f00),
+      onPrimaryFixed: Color(0xffffffff),
+      primaryFixedDim: Color(0xff4c2b00),
+      onPrimaryFixedVariant: Color(0xffffffff),
+      secondaryFixed: Color(0xff5b452e),
+      onSecondaryFixed: Color(0xffffffff),
+      secondaryFixedDim: Color(0xff422f1a),
+      onSecondaryFixedVariant: Color(0xffffffff),
+      tertiaryFixed: Color(0xff424d27),
+      onTertiaryFixed: Color(0xffffffff),
+      tertiaryFixedDim: Color(0xff2c3613),
+      onTertiaryFixedVariant: Color(0xffffffff),
+      surfaceDim: Color(0xffc4b6ac),
+      surfaceBright: Color(0xfffff8f4),
+      surfaceContainerLowest: Color(0xffffffff),
+      surfaceContainerLow: Color(0xfffdeee3),
+      surfaceContainer: Color(0xffeee0d5),
+      surfaceContainerHigh: Color(0xffe0d2c7),
+      surfaceContainerHighest: Color(0xffd2c4b9),
+    );
+  }
+
+  static ColorScheme darkScheme() {
+    return const ColorScheme(
+      brightness: Brightness.dark,
+      primary: Color(0xfffbb974),
+      surfaceTint: Color(0xfffbb974),
+      onPrimary: Color(0xff492900),
+      primaryContainer: Color(0xff683d00),
+      onPrimaryContainer: Color(0xffffdcbc),
+      secondary: Color(0xffe1c1a3),
+      onSecondary: Color(0xff402c18),
+      secondaryContainer: Color(0xff58432c),
+      onSecondaryContainer: Color(0xfffeddbe),
+      tertiary: Color(0xffbecc9b),
+      onTertiary: Color(0xff293411),
+      tertiaryContainer: Color(0xff3f4b25),
+      onTertiaryContainer: Color(0xffdae8b5),
+      error: Color(0xffffb4ab),
+      onError: Color(0xff690005),
+      errorContainer: Color(0xff93000a),
+      onErrorContainer: Color(0xffffdad6),
+      surface: Color(0xff19120c),
+      onSurface: Color(0xffeee0d5),
+      onSurfaceVariant: Color(0xffd5c3b5),
+      outline: Color(0xff9d8e81),
+      outlineVariant: Color(0xff50453a),
+      shadow: Color(0xff000000),
+      scrim: Color(0xff000000),
+      inverseSurface: Color(0xffeee0d5),
+      inversePrimary: Color(0xff845317),
+      primaryFixed: Color(0xffffdcbc),
+      onPrimaryFixed: Color(0xff2c1700),
+      primaryFixedDim: Color(0xfffbb974),
+      onPrimaryFixedVariant: Color(0xff683d00),
+      secondaryFixed: Color(0xfffeddbe),
+      onSecondaryFixed: Color(0xff291805),
+      secondaryFixedDim: Color(0xffe1c1a3),
+      onSecondaryFixedVariant: Color(0xff58432c),
+      tertiaryFixed: Color(0xffdae8b5),
+      onTertiaryFixed: Color(0xff151f01),
+      tertiaryFixedDim: Color(0xffbecc9b),
+      onTertiaryFixedVariant: Color(0xff3f4b25),
+      surfaceDim: Color(0xff19120c),
+      surfaceBright: Color(0xff403830),
+      surfaceContainerLowest: Color(0xff130d07),
+      surfaceContainerLow: Color(0xff211a14),
+      surfaceContainer: Color(0xff251e17),
+      surfaceContainerHigh: Color(0xff302921),
+      surfaceContainerHighest: Color(0xff3c332c),
+    );
+  }
+
+  static ColorScheme darkMediumContrastScheme() {
+    return const ColorScheme(
+      brightness: Brightness.dark,
+      primary: Color(0xffffd5ac),
+      surfaceTint: Color(0xfffbb974),
+      onPrimary: Color(0xff3a1f00),
+      primaryContainer: Color(0xffbf8544),
+      onPrimaryContainer: Color(0xff000000),
+      secondary: Color(0xfff8d7b8),
+      onSecondary: Color(0xff34220e),
+      secondaryContainer: Color(0xffa88c71),
+      onSecondaryContainer: Color(0xff000000),
+      tertiary: Color(0xffd4e2b0),
+      onTertiary: Color(0xff1f2907),
+      tertiaryContainer: Color(0xff899669),
+      onTertiaryContainer: Color(0xff000000),
+      error: Color(0xffffd2cc),
+      onError: Color(0xff540003),
+      errorContainer: Color(0xffff5449),
+      onErrorContainer: Color(0xff000000),
+      surface: Color(0xff19120c),
+      onSurface: Color(0xffffffff),
+      onSurfaceVariant: Color(0xffebd9ca),
+      outline: Color(0xffbfafa1),
+      outlineVariant: Color(0xff9d8e81),
+      shadow: Color(0xff000000),
+      scrim: Color(0xff000000),
+      inverseSurface: Color(0xffeee0d5),
+      inversePrimary: Color(0xff6a3e00),
+      primaryFixed: Color(0xffffdcbc),
+      onPrimaryFixed: Color(0xff1d0d00),
+      primaryFixedDim: Color(0xfffbb974),
+      onPrimaryFixedVariant: Color(0xff512e00),
+      secondaryFixed: Color(0xfffeddbe),
+      onSecondaryFixed: Color(0xff1d0e01),
+      secondaryFixedDim: Color(0xffe1c1a3),
+      onSecondaryFixedVariant: Color(0xff46321d),
+      tertiaryFixed: Color(0xffdae8b5),
+      onTertiaryFixed: Color(0xff0c1400),
+      tertiaryFixedDim: Color(0xffbecc9b),
+      onTertiaryFixedVariant: Color(0xff2f3a16),
+      surfaceDim: Color(0xff19120c),
+      surfaceBright: Color(0xff4c433b),
+      surfaceContainerLowest: Color(0xff0c0603),
+      surfaceContainerLow: Color(0xff231c16),
+      surfaceContainer: Color(0xff2e261f),
+      surfaceContainerHigh: Color(0xff39312a),
+      surfaceContainerHighest: Color(0xff453c34),
+    );
+  }
+
+  static ColorScheme darkHighContrastScheme() {
+    return const ColorScheme(
+      brightness: Brightness.dark,
+      primary: Color(0xffffedde),
+      surfaceTint: Color(0xfffbb974),
+      onPrimary: Color(0xff000000),
+      primaryContainer: Color(0xfff7b670),
+      onPrimaryContainer: Color(0xff150800),
+      secondary: Color(0xffffedde),
+      onSecondary: Color(0xff000000),
+      secondaryContainer: Color(0xffdcbda0),
+      onSecondaryContainer: Color(0xff150800),
+      tertiary: Color(0xffe8f6c2),
+      onTertiary: Color(0xff000000),
+      tertiaryContainer: Color(0xffbbc897),
+      onTertiaryContainer: Color(0xff070d00),
+      error: Color(0xffffece9),
+      onError: Color(0xff000000),
+      errorContainer: Color(0xffffaea4),
+      onErrorContainer: Color(0xff220001),
+      surface: Color(0xff19120c),
+      onSurface: Color(0xffffffff),
+      onSurfaceVariant: Color(0xffffffff),
+      outline: Color(0xffffedde),
+      outlineVariant: Color(0xffd1c0b1),
+      shadow: Color(0xff000000),
+      scrim: Color(0xff000000),
+      inverseSurface: Color(0xffeee0d5),
+      inversePrimary: Color(0xff6a3e00),
+      primaryFixed: Color(0xffffdcbc),
+      onPrimaryFixed: Color(0xff000000),
+      primaryFixedDim: Color(0xfffbb974),
+      onPrimaryFixedVariant: Color(0xff1d0d00),
+      secondaryFixed: Color(0xfffeddbe),
+      onSecondaryFixed: Color(0xff000000),
+      secondaryFixedDim: Color(0xffe1c1a3),
+      onSecondaryFixedVariant: Color(0xff1d0e01),
+      tertiaryFixed: Color(0xffdae8b5),
+      onTertiaryFixed: Color(0xff000000),
+      tertiaryFixedDim: Color(0xffbecc9b),
+      onTertiaryFixedVariant: Color(0xff0c1400),
+      surfaceDim: Color(0xff19120c),
+      surfaceBright: Color(0xff584e46),
+      surfaceContainerLowest: Color(0xff000000),
+      surfaceContainerLow: Color(0xff251e17),
+      surfaceContainer: Color(0xff372f28),
+      surfaceContainerHigh: Color(0xff423a32),
+      surfaceContainerHighest: Color(0xff4e453d),
+    );
+  }
+}
